+++ src/test/java/blackjack/model/betting/BettingAmountTest.java
@@ -1,4 +1,4 @@
-package blackjack.model.participant;
+package blackjack.model.betting;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
